+++ Products/CMFPlone/resources/browser/scripts.py
@@ -111,7 +111,7 @@ class ScriptsView(ResourceView):
         """The requirejs scripts, the ones that are not resources are loaded on
         configjs.py
         """
-        if self.development:
+        if self.development or not self.production_path:
             result = self.default_resources()
             result.extend(self.ordered_bundles_result())
         else: